+++ Tests/Editor/Scripts/Timeline/TimelineEditorUtilityTest.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections;
+using System.Collections.Generic;
 using System.IO;
 using NUnit.Framework;
 using Unity.FilmInternalUtilities.Editor;
@@ -26,7 +27,7 @@ internal class TimelineEditorUtilityTest {
                 
     [Test]
     public void CreateAndDestroyTimelineAssets() {
-        PlayableDirector director          = CreateDirectorWithTimelineAsset(TIMELINE_ASSET_PATH);
+        PlayableDirector director = CreateDirectorWithTimelineAsset(TIMELINE_ASSET_PATH,out TimelineAsset _);
         string           timelineAssetPath = AssetDatabase.GetAssetPath(director.playableAsset);
         TimelineEditorUtility.DestroyAssets(director.playableAsset);
         Assert.IsFalse(File.Exists(timelineAssetPath));
@@ -37,15 +38,18 @@ internal class TimelineEditorUtilityTest {
 
     [UnityTest]
     public IEnumerator CreateClip() {
-        PlayableDirector director      = CreateDirectorWithTimelineAsset(TIMELINE_ASSET_PATH);
-        TimelineAsset    timelineAsset = director.playableAsset as TimelineAsset;
-        Assert.IsNotNull(timelineAsset);
+        PlayableDirector director = CreateDirectorWithTimelineAsset(TIMELINE_ASSET_PATH,out TimelineAsset timelineAsset);
         yield return EditorTestsUtility.WaitForFrames(1);
         
         TimelineClip clip = TimelineEditorUtility.CreateTrackAndClip<DummyTimelineTrack, DummyTimelinePlayableAsset>(
             timelineAsset, "FirstTrack");
         VerifyClip(clip);
-
+        TimelineEditorUtility.SelectDirectorInTimelineWindow(director); //trigger the TimelineWindow's update etc.
+        yield return EditorTestsUtility.WaitForFrames(3);
+        
+        DummyTimelineClipData clipData = clip.GetClipData<DummyTimelineClipData>();
+        Assert.IsNotNull(clipData);
+        
         TimelineEditorUtility.DestroyAssets(clip); //Cleanup
     }
 
@@ -53,9 +57,7 @@ internal class TimelineEditorUtilityTest {
     
     [UnityTest]
     public IEnumerator ShowClipInInspector() {
-        PlayableDirector director      = CreateDirectorWithTimelineAsset(TIMELINE_ASSET_PATH);
-        TimelineAsset    timelineAsset = director.playableAsset as TimelineAsset;
-        Assert.IsNotNull(timelineAsset);
+        CreateDirectorWithTimelineAsset(TIMELINE_ASSET_PATH,out TimelineAsset timelineAsset);
         yield return EditorTestsUtility.WaitForFrames(1);
         
         TrackAsset track = timelineAsset.CreateTrack<DummyTimelineTrack>(null, "FooTrack");        
@@ -72,15 +74,64 @@ internal class TimelineEditorUtilityTest {
         TimelineEditorUtility.DestroyAssets(clip); //Cleanup
 
     }
+
+//----------------------------------------------------------------------------------------------------------------------
+    
+    [UnityTest]
+    public IEnumerator DetectActiveClip() {
+        PlayableDirector director = CreateDirectorWithTimelineAsset(TIMELINE_ASSET_PATH,out TimelineAsset timelineAsset);
+        TimelineEditorUtility.SelectDirectorInTimelineWindow(director); //trigger the TimelineWindow's update etc.
+        yield return EditorTestsUtility.WaitForFrames(1);
+        
+        TrackAsset         track     = timelineAsset.CreateTrack<DummyTimelineTrack>(null, "FooTrack");
+        List<TimelineClip> clips     = new List<TimelineClip>();
+        const int          NUM_CLIPS = 3;
+
+        double nextClipStart = 0;
+        for (int i = 0; i < NUM_CLIPS; ++i) {
+            TimelineClip curClip = TimelineEditorReflection.CreateClipOnTrack(typeof(DummyTimelinePlayableAsset), track, 0);
+            curClip.asset.name =  $"Clip Asset {i}";
+            curClip.start      =  nextClipStart;
+            nextClipStart      += curClip.duration;
+            clips.Add(curClip);
+            
+            yield return EditorTestsUtility.WaitForFrames(1);
+        }
+
+        //Check that all clips have been created successfully
+        List<TimelineClip>    trackClips        = new List<TimelineClip>(track.GetClips());
+        HashSet<TimelineClip> trackClipsToCheck = new HashSet<TimelineClip>(trackClips);
+        for (int i = 0; i < NUM_CLIPS; ++i) {
+            Assert.IsTrue(trackClipsToCheck.Contains(clips[i]));
+            trackClipsToCheck.Remove(clips[i]);
+        }
+        NUnit.Framework.Assert.Zero(trackClipsToCheck.Count);
+        yield return EditorTestsUtility.WaitForFrames(3);
+        
+        //Ensure the active clip can be detected
+        for (int i = 0; i < NUM_CLIPS; ++i) {
+            TimelineClip curClip = clips[i];
+            double time = curClip.start + curClip.duration * 0.5f;
+            TimelineUtility.GetActiveTimelineClipInto(trackClips, time, out TimelineClip detectedClip, out TimelineAsset _);
+            Assert.AreEqual(curClip, detectedClip);
+        }
+       
+        yield return EditorTestsUtility.WaitForFrames(1);
+                   
+        TimelineEditorUtility.DestroyAssets(timelineAsset); //Cleanup
+
+    }
     
 //----------------------------------------------------------------------------------------------------------------------
 
 
-    private static PlayableDirector CreateDirectorWithTimelineAsset(string candidatePath) {
-        string timelineAssetPath = AssetDatabase.GenerateUniqueAssetPath(candidatePath);
-        
-        PlayableDirector director   = new GameObject("Director").AddComponent<PlayableDirector>();
-        TimelineAsset timelineAsset = TimelineEditorUtility.CreateAsset(timelineAssetPath);
+    private static PlayableDirector CreateDirectorWithTimelineAsset(string candidatePath, 
+        out TimelineAsset timelineAsset) 
+    {
+        string           timelineAssetPath = AssetDatabase.GenerateUniqueAssetPath(candidatePath);        
+        PlayableDirector director          = new GameObject("Director").AddComponent<PlayableDirector>();
+        timelineAsset = TimelineEditorUtility.CreateAsset(timelineAssetPath);
+        Assert.IsNotNull(timelineAsset);
 
         director.playableAsset = timelineAsset; 
         Assert.IsTrue(File.Exists(timelineAssetPath));