+++ Tests/Runtime/Scripts/Dummy/DummyScriptableObject.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 namespace Unity.FilmInternalUtilities.Tests {
 
 [Serializable]
-internal class DummyScriptableObject : ScriptableObject{
+internal class DummyScriptableObject : ScriptableObject {
+    [SerializeField] internal ExposedReference<GameObject> exposedGameObject;
 }
-
 
 } //end namespace